i156 Move JobContextSetupListener to a public package
... since it's supposed to be extended.
